refactor[#107] : getPopularArticleList 오타 수정
